backend/requirements.txt
# VAIDA Backend - Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.18
httpx==0.28.1
openai==1.58.1
reportlab==4.2.5
jinja2==3.1.4

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0

# Production extras
# asyncpg==0.30.0       # Uncomment for PostgreSQL
# psycopg2-binary==2.9.10  # Uncomment for PostgreSQL
# sentry-sdk[fastapi]==2.19.2
# prometheus-fastapi-instrumentator==7.0.2
